+++ app.js
@@ -811,6 +811,7 @@ import { renderers as RV1443 } from './views/group-v1443.js';
 import { renderers as RV1444 } from './views/group-v1444.js';
 import { renderers as RV1445 } from './views/group-v1445.js';
 import { renderers as RV1446 } from './views/group-v1446.js';
+import { renderers as RV1447 } from './views/group-v1447.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -899,7 +900,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -2033,6 +2034,7 @@ const UTILITIES = [
   { id: 'frykman-classification', name: 'Frykman Classification (distal radius)',             group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'mirels-score',           name: 'Mirels Score (impending pathologic fracture)',       group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'kellgren-lawrence',      name: 'Kellgren-Lawrence Osteoarthritis Grade',            group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'cpak', name: 'CPAK Classification (Knee Alignment Phenotype)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'outerbridge-cartilage',  name: 'Outerbridge Cartilage Grade (0-IV)',                group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'isakos-meniscal', name: 'ISAKOS Classification (Meniscal Tears)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'icrs-cartilage',         name: 'ICRS Cartilage Lesion Grade (0-4)',                 group: 'G', audiences: ['clinicians', 'educators'], clinical: true },